Rename "EmailProcess" to IssueEmail Log event details when recover occurs in Graph
